[FIX] Add date file using datetime module
